move sgf projects fetching to service
