+++ pages/about.js
@@ -3,7 +3,7 @@ import groupBy from 'lodash/groupBy';
 import PageWrapper from 'layouts/wrappers/page';
 import About from 'layouts/about';
 
-import { fetchAboutProjects, fetchImpactProjects } from 'services/projects';
+import { getUseCaseProjects, getImpactProjects } from 'services/projects';
 
 const AboutPage = (props) => (
   <PageWrapper
@@ -15,7 +15,7 @@ const AboutPage = (props) => (
 );
 
 export const getStaticProps = async () => {
-  const projects = await fetchAboutProjects();
+  const projects = await getUseCaseProjects();
   const groupedProjects = groupBy(projects, 'category');
   const categories = [
     { label: 'All', count: projects.length },
@@ -29,7 +29,7 @@ export const getStaticProps = async () => {
     All: projects,
   };
 
-  const impactProjects = await fetchImpactProjects();
+  const impactProjects = await getImpactProjects();
 
   return {
     props: {
