+++ pages/subscribe.js
@@ -3,9 +3,9 @@ import Subscribe from 'layouts/subscribe';
 
 const SubscribePage = () => (
   <PageLayout
-    noIndex
     title="Stay Updated on the World's Forests | Global Forest Watch"
     description="Subscribe to monthly GFW newsletters and updates based on your interests."
+    noIndex
   >
     <Subscribe />
   </PageLayout>
