+++ services/country.js
@@ -15,7 +15,7 @@ const SQL_QUERIES = {
   getRanking:
     "WITH mytable AS (SELECT fao.iso, fao.name, fao.forest_primary, fao.extent forest_extent, a.land as area_ha FROM gfw2_countries as fao INNER JOIN umd_nat_staging as a ON fao.iso = a.iso WHERE fao.forest_primary > 0 AND a.year = 2001 AND a.thresh = 30), rank AS ( SELECT forest_extent * (forest_primary/100)/area_ha * 100 as percent_primary ,iso from mytable ORDER BY percent_primary DESC), item as (select percent_primary from rank where iso = '{adm0}') select count(*) as rank from rank WHERE percent_primary > (select percent_primary from item )",
   getCountriesLatLng:
-    'SELECT latitude_average, longitude_average, alpha_3_code as iso FROM country_list_iso_3166_codes_latitude_longitude'
+    'SELECT latitude_average, longitude_average, alpha_3_code as iso FROM country_list_iso_3166_codes_latitude_longitude',
 };
 
 export const getCountriesProvider = () => {
@@ -40,14 +40,14 @@ export const getSubRegionsProvider = (adm0, adm1, token) => {
   return cartoRequest.get(url, { cancelToken: token });
 };
 
-export const getCountryLinksProvider = token => {
+export const getCountryLinksProvider = (token) => {
   const url = `/sql?q=${SQL_QUERIES.getCountryLinks}`;
   return cartoRequest.get(url, { cancelToken: token });
 };
 
 export const getCountriesLatLng = () => {
   const url = `/sql?q=${SQL_QUERIES.getCountriesLatLng}`;
-  return cartoRequest.get(url);
+  return cartoRequest.get(url).then((response) => response?.data?.rows);
 };
 
 export const getRanking = ({ adm0, token }) => {
